Reorganize research models
